% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_chole.R
\docType{data}
\name{nhanes_chole}
\alias{nhanes_chole}
\title{Le jeu de donnée simplifié de nhanes sur Cholesterol.}
\format{Un jeu de donnée contient 8327 individus et ? variables:
\describe{
  \item{SEQN}{les individus}
  \item{y}{la varible de sortie ici concerne le cholesterol}
  \item{_dr1tot}{les variables reliées aux données de nutrition}
  ...
}}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2015/}
}
\usage{
nhanes_chole
}
\description{
Le jeu de donnée simplifié de nhanes sur Cholesterol.
}
\keyword{datasets}
