% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_hyper.R
\docType{data}
\name{nhanes_hyper}
\alias{nhanes_hyper}
\title{Le jeu de donnée simplifié de nhanes sur hypertension.}
\format{Un jeu de donnée contient 8327 individus et ? variables:
\describe{
  \item{SEQN}{les individus}
  \item{y}{la varible de sortie ici concerne le cholesterol'hypertension}
  \item{_dr1tot}{les variables reliées aux données de nutrition}
  ...
}}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2015/}
}
\usage{
nhanes_hyper
}
\description{
Le jeu de donnée simplifié de nhanes sur hypertension.
}
\keyword{datasets}
