% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_obesite.R
\docType{data}
\name{nhanes_obesite}
\alias{nhanes_obesite}
\title{Le jeu de donnée simplifié de nhanes sur l'obésité.}
\format{Un jeu de donnée contient 8327 individus et 75 variables:
\describe{
  \item{SEQN}{les individus}
  \item{y}{la varible de sortie ici concerne l'obésité}
  \item{_dr1tot}{les variables reliées aux données de nutrition}
  ...
}}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2015/}
}
\usage{
nhanes_obesite
}
\description{
Le jeu de donnée simplifié de nhanes sur l'obésité.
}
\keyword{datasets}
