% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transco.R
\docType{data}
\name{transco}
\alias{transco}
\title{Le jeu de donnée pour transcodifier du code vers libellé.}
\format{Un jeu de donnée contient 8465 individus et 9 variables:
\describe{
  \item{Code.Or.Value}{les codes à transformer}
  \item{Value.Desciption}{La desciption des codes}
  ...
}}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2015/}
}
\usage{
transco
}
\description{
Le jeu de donnée pour transcodifier du code vers libellé.
}
\keyword{datasets}
